requirements.txt
# data analysis libraries
numpy~=1.20.1
scipy
pandas~=1.2.2
pandas-datareader

# technical analysis, indicator libraries
ta~=0.7.0
ta-lib
bta-lib
pandas-ta
tulipy~=0.4.0

# database connection, ORM
psycopg2
sqlalchemy

# broker API libraries
alpaca-trade-api
python-binance
tda-api
ib_insync
robin-stocks

# data providers
intrinio-sdk
polygon-api-client
iexfinance
yfinance~=0.1.55
quandl
alpha-vantage
sec-edgar-downloader

# visualization
matplotlib
plotly~=4.14.3
dash
mplfinance
jupyterlab
pillow

# backtesting
backtrader~=1.9.76.123
pyalgotrade
bt

# portfolio, performance analysis
pyfolio
finquant

# web frameworks and extensions
fastapi~=0.63.0
flask
flask-cors
flask-login
flask-sqlalchemy
itsdangerous
aiofiles
chalice

# servers, task queue
uvicorn
gunicorn
celery

# networking
requests~=2.25.1
boto3
urllib3
websocket-client
websockets

# other data stores, formats
redis
h5py

# utilities
beautifulsoup4
pendulum
click
passlib

# machine learning
tensorflow
scikit-learn
keras
torch
opencv-python
pytz~=2021.1